+++ packages/components/src/components/radio/radio.types.ts
@@ -1,0 +1,4 @@
+export type RadioChangeDetail = {
+  checked: boolean;
+  value: string;
+};