+++ packages/components/src/index.ts
@@ -1,2 +1,3 @@
 export * from './components/button/index.js';
 export * from './components/checkbox/index.js';
+export * from './components/radio/index.js';